Island: Move authentication_service.utils to flask_resources
